requirements.txt
openai==2.13.0
langchain==0.3.27
langchain-community==0.3.31
langchain-openai==0.3.35
langchain-chroma==0.2.6
chromadb==1.3.7
pydantic==2.12.5
langfuse==3.7.0
python-dotenv==1.2.1
tiktoken==0.12.0
pypdf==6.4.2
fastapi==0.125.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
python-multipart==0.0.20
python-json-logger==2.0.7
